Update IBRES
This questionnaire aims to gauge someone's proneness to aversive or symbolic racism. At the end of the questionnaire, it will tell you if:

1. "you tend to have a more accepting or positive view in (aversive or symbolic racism)"

2. "You may have concerns or a less accepting view in (aversive or symbolic racism)", or 

3. "Your responses in (aversive or symbolic racism)' were neutral."

It's important to note that accurately gauging someone's propensity for aversive or symbolic racism can be complex. This test does not gauge on whether you are or aren't racist, but it instead gives insight to what you may be more prone to be susceptible to given certain situations. This test is about awareness, and is not to be used to judge, bully, or harm anyone. Racism is learned and can be unlearned. Education and informing people is key to lessening the hold racism has on them. 

This is recommended to be taken privately to avoid any potential judgement and for honesty. If this test is taken with others, I am not responsible for how others react.